move things around to make space for DecodeAsType and proc macro crate
