+++ .rustfmt.toml
@@ -1,4 +1,5 @@
-hard_tabs = true
+hard_tabs = false
+tab_spaces = 4
 max_width = 100
 use_small_heuristics = "Max"
 edition = "2021"
