+++ Cargo.toml
@@ -1,25 +1,4 @@
-[package]
-name = "scale-decode"
-version = "0.4.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2021"
-
-license = "Apache-2.0"
-readme = "README.md"
-repository = "https://github.com/paritytech/scale-decode"
-documentation = "https://docs.rs/scale-decode"
-homepage = "https://www.parity.io/"
-description = "Decode SCALE encoded bytes into arbitrary types at runtime"
-keywords = ["parity", "scale", "decoding"]
-include = ["Cargo.toml", "src/**/*.rs", "README.md", "LICENSE"]
-
-[dependencies]
-scale-info = { version = "2.0.0", default-features = false, features = ["bit-vec", "std"] }
-codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive", "full"] }
-thiserror = "1.0.24"
-scale-bits = { version = "0.3.0" }
-
-[dev-dependencies]
-scale-info = { version = "2.0.0", default-features = false, features = ["bit-vec", "std", "derive"] }
-codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive", "full", "bit-vec"] }
-bitvec = "1"
+[workspace]
+members = [
+    "scale-decode",
+]
