+++ scale-decode/src/visitor/compact.rs
@@ -17,52 +17,52 @@ use super::TypeId;
 
 /// This represents a compact encoded type.
 pub struct Compact<'b, 'c, T> {
-	val: T,
-	locations: &'c [CompactLocation<'b>],
+    val: T,
+    locations: &'c [CompactLocation<'b>],
 }
 
 impl<'b, 'c, T: Copy> Compact<'b, 'c, T> {
-	pub(crate) fn new(val: T, locations: &'c [CompactLocation<'b>]) -> Compact<'b, 'c, T> {
-		Compact { val, locations }
-	}
-	/// Return the value that was compact-encoded
-	pub fn value(&self) -> T {
-		self.val
-	}
-	/// Compact values can be nested inside named or unnamed fields in structs.
-	/// This provides back a slice of
-	pub fn locations(&self) -> &'c [CompactLocation<'b>] {
-		self.locations
-	}
+    pub(crate) fn new(val: T, locations: &'c [CompactLocation<'b>]) -> Compact<'b, 'c, T> {
+        Compact { val, locations }
+    }
+    /// Return the value that was compact-encoded
+    pub fn value(&self) -> T {
+        self.val
+    }
+    /// Compact values can be nested inside named or unnamed fields in structs.
+    /// This provides back a slice of
+    pub fn locations(&self) -> &'c [CompactLocation<'b>] {
+        self.locations
+    }
 }
 
 /// A pointer to what the compact value is contained within.
 #[derive(Clone, Copy, Debug)]
 pub enum CompactLocation<'b> {
-	/// We're in an unnamed composite (struct) with the type ID given.
-	UnnamedComposite(TypeId),
-	/// We're in a named composite (struct) with the type ID given, and the compact
-	/// value lives inside the field with the given name.
-	NamedComposite(TypeId, &'b str),
-	/// We're at a primitive type with the type ID given; the compact value itself.
-	Primitive(TypeId),
+    /// We're in an unnamed composite (struct) with the type ID given.
+    UnnamedComposite(TypeId),
+    /// We're in a named composite (struct) with the type ID given, and the compact
+    /// value lives inside the field with the given name.
+    NamedComposite(TypeId, &'b str),
+    /// We're at a primitive type with the type ID given; the compact value itself.
+    Primitive(TypeId),
 }
 
 impl<'b> CompactLocation<'b> {
-	/// Return the Primitive type of this location, if one exists.
-	pub fn as_primitive(self) -> Option<TypeId> {
-		match self {
-			CompactLocation::Primitive(t) => Some(t),
-			_ => None,
-		}
-	}
+    /// Return the Primitive type of this location, if one exists.
+    pub fn as_primitive(self) -> Option<TypeId> {
+        match self {
+            CompactLocation::Primitive(t) => Some(t),
+            _ => None,
+        }
+    }
 }
 
 // Default values for locations are never handed back, but they are
 // stored on the StackArray in the "unused" positions. We could avoid needing
 // this with some unsafe code.
 impl<'a> Default for CompactLocation<'a> {
-	fn default() -> Self {
-		CompactLocation::Primitive(TypeId::default())
-	}
+    fn default() -> Self {
+        CompactLocation::Primitive(TypeId::default())
+    }
 }
